fix images and main.ts
